% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRows.R
\name{addRows}
\alias{addRows}
\title{Add rows to a dataset (V2)}
\usage{
addRows(apiKey, datasetId, dataframeName, chunkSize = 1500, printSample = F,
  nullArrays = F)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{datasetId}{The dataset ID, in quotes. It can be found by using the getDatasets function, or by navigating to the dataset in Numetric, and selecting the string after the last forward slash.}

\item{dataframeName}{The name of the dataframe, without quotes.}

\item{chunkSize}{Defaults to 1,500 rows.}

\item{printSample}{Defaults to False. If set to True, then it will return a sample of the data that is being indexed to verify that it's correctly being converted to JSON.}

\item{nullArrays}{Defaults to False. Set to true if your data has arrays within a cell, and if some of those arrays have NULL values.}
}
\value{
Returns the status (e.g., 200, 401).
}
\description{
Adds data to a new or existing dataset. If the primary key already exists in the indexed data, then the old data will be replaced with the new data.

If the primary key doesn't already exist in the dataset, then the new data will be added on to the existing data.

This function uses the V2 version of the API.
}

