% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucketQuery.R
\name{bucketQuery}
\alias{bucketQuery}
\title{Bucket query a Numetric dataset (V2)}
\usage{
bucketQuery(apiKey, datasetId, bucketVar)
}
\arguments{
\item{apiKey}{The API key, in quotes. You can find the API key in the settings after logging into Numetric}

\item{datasetId}{The dataset ID, in quotes. The dataset ID can be found by using the getDatasets function, or by navigating to the dataset in Numetric, and selecting the string after the last forward slash.}

\item{bucketVar}{The name of the dataset column, in quotes, that will be used to bucket the data.}
}
\value{
Returns a dataframe.
}
\description{
Returns a dataframe with a list of bucket variable values, and the number of observations for each bucket.

 This is especially useful when you have a dataset larger than 10,000 rows because it allows you to create a loop through the data getting only the data for each bucket.

This function uses the V2 version of the API.
}

