% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDashboard.R
\name{createDashboard}
\alias{createDashboard}
\title{Creates a new dashboard in Numetric (V2)}
\usage{
createDashboard(apiKey, name = "None", category = "None", content)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{name}{The name of the dashboard, in quotes, in Numetric. Defaults to "None".}

\item{content}{JSON formatted string of text. See https://nrl.readme.io/docs for more information.}
}
\description{
Creates a new dashboard in Numetric.

This function uses the V2 version of the API.
}
