% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTable.R
\name{createTable}
\alias{createTable}
\title{Create a new table (V3)}
\usage{
createTable(apiKey, numetricName, dataframeName, category = "New Data",
  primaryKey, geoshapes = "", geopoints = "", boolean = "",
  everyone = "false")
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{numetricName}{The name, in quotes, given to the dataset in the Numetric Warehouse.}

\item{dataframeName}{The name, without quotes, of the dataframe that will be used to create the dataset in Numetric.}

\item{category}{The category, in quotes, given to the dataset in the Numetric Warehouse.}

\item{geoshapes}{The column name, or vector of column names, in quotes, that will be stored as a geoShape (point on a map). This needs to be in the format, "lat,long".}

\item{geopoints}{The column name, or vector of column names, in quotes, that will be stored as a geoPoint (for heatmaps). This needs to be in the format, "lat,long".}

\item{boolean}{The column name, or vector of column names, in quotes, that will be stored as True/False. True = 1, False = 0.}

\item{everyone}{Defaults to "false". If set to "true", then it will allow everyone in the org to see the dataset.}

\item{id}{The name of the column, in quotes, that will be used as the primary key. Each row should have a unique id, otherwise, the last row indexed will be the one saved.}

\item{autocompletes}{The column name, or vector of column names, in quotes, that will autocomplete in the Numetric quick search bar.}
}
\value{
Numetric Id
}
\description{
Create a new table in a Numetric Org.

This function uses the V3 version of the API.
}
