% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteAllRowsV3.R
\name{deleteAllRowsV3}
\alias{deleteAllRowsV3}
\title{Delete all Rows of a Table (V3)}
\usage{
deleteAllRowsV3(apiKey, tableId)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{tableId}{The id of the table. You can find this by using the getTables function, or by navigating to the table, and then selecting the string after the last forward slash of the url.}
}
\description{
RDeletes all the rows in a table. This function uses the V3 version of the API.
}
