% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDashboard.R
\name{deleteDashboard}
\alias{deleteDashboard}
\title{Deletes a dashboard (V2)}
\usage{
deleteDashboard(apiKey, dashboardId)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{dashboardId}{The dashboard ID, in quotes. You can find the dashboard ID by either (1) using the getDashboards function, or (2) navigating to the dashboard, and then selecting the string after the last forward slash.}
}
\description{
Deletes a dashboard using the API.

This function uses the V2 version of the API.
}
