% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDataset.R
\name{deleteDataset}
\alias{deleteDataset}
\title{Deletes a dataset (V2)}
\usage{
deleteDataset(apiKey, datasetId)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{datasetId}{The dataset ID, in quotes. You can find the dataset ID by either (1) using the getDatasets function, or (2) navigating to the dataset, and then selecting the string after the last forward slash.}
}
\description{
Deletes a dataset using the API.

This function uses the V2 version of the API.
}

