% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSomeRows.R
\name{deleteSomeRows}
\alias{deleteSomeRows}
\title{Deletes specific rows from a dataset (V2)}
\usage{
deleteSomeRows(apiKey, datasetId, rows)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{datasetId}{The dataset ID, in quotes. You can find the dataset ID by either (1) using the getDatasets function, or (2) navigating to the dataset, and then selecting the string after the last forward slash.}

\item{rows}{A vector of rows containing the primary keys of the rows to be deleted.}
}
\description{
Deletes rows of a dataset based on a vector of primary keys.

This function uses the V2 version of the API.
}

