% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTable.R
\name{deleteTable}
\alias{deleteTable}
\title{Deletes a table (V3)}
\usage{
deleteTable(apiKey, tableId)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{tableId}{The table ID, in quotes. You can find the table ID by either (1) using the getTables function, or (2) navigating to the table, and then selecting the string after the last forward slash.}
}
\description{
Deletes a table using the API.

This function uses the V3 version of the API.
}
