% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllRowsV3.R
\name{getAllRowsV3}
\alias{getAllRowsV3}
\title{Get all rows of data from a table (V3)}
\usage{
getAllRowsV3(apiKey, datasetId)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{datasetId}{The id of the dataset. This can be found by navigating to the dataset in Numetric, and then referring to the string of characters after the last forward slash in the URL.}
}
\value{
Numetric Id
}
\description{
Returns every single row of a V3 table.

This function uses the V3 version of the API.
}
