% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDashboards.R
\name{getDashboards}
\alias{getDashboards}
\title{Get a meta-data for all dashboards in an org (V2)}
\usage{
getDashboards(apiKey)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}
}
\value{
Returns a dataframe.
}
\description{
Returns a dataframe with meta-data about all of the dashboards in an org.

This function uses the V2 version of the API.
}

