% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasets.R
\name{getDatasets}
\alias{getDatasets}
\title{Get a table of datasets (V2)}
\usage{
getDatasets(apiKey)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}
}
\description{
Returns a table with meta-data about all the tables. This function uses the V2 version of the API.
}
