% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetStatus.R
\name{getDatasetStatus}
\alias{getDatasetStatus}
\title{Get the status of a dataset (V2)}
\usage{
getDatasetStatus(apiKey, datasetId)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{datasetId}{The dataset ID can be found by using the getDatasets function, or by navigating to the dataset in Numetric, and selecting the string after the last forward slash.}
}
\value{
Returns a dataframe.
}
\description{
Find out if the index is complete.  This function uses the V2 version of the API.
}
