% Please edit documentation in R/getDatasetsV3.R
\name{getDatasetsV3}
\alias{getDatasetsV3}
\title{Get a summary list of tables available to the API user (V3)}
\usage{
getDatasetsV3(apiKey)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}
}
\description{
Returns a table with meta-data about all the datasets. This function uses the V3 version of the API.
}
