% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilteredRows.R
\name{getFilteredRows}
\alias{getFilteredRows}
\title{Get up to 10,000 rows of a filtered dataset (V2)}
\usage{
getFilteredRows(apiKey, datasetId, filterType = "term", filterField = "",
  filterValue = "", startRange = "", endRange = "")
}
\arguments{
\item{apiKey}{The API key, in quotes. You can find the API key in the settings after logging into Numetric}

\item{filterType}{The type of filter (term or range), in quotes. Defaults to "term", but can also be "range" for a numeric or date range.}

\item{filterField}{The name of the column, in quotes, to be used as the filter.}

\item{filterValue}{If the filterType == term, then this should be the value, in quotes, by which to filter the data.}

\item{startRange}{If the filterType == range, then this is a minimum numeric value or date, in quotes. If it's a date, then it should be in the format of "2017-05-30T00.00.00.000".}

\item{endRange}{If the filterType == range, then this is a maximum numeric value or date, in quotes. If it's a date, then it should be in the format of "2017-05-30T00.00.00.000".}

\item{bucketVar}{The name of the dataset column, in quotes, that will be used to bucket the data.}
}
\value{
Returns a dataframe.
}
\description{
Returns a dataframe with up to 10,000 rows of a dataset after applying a term or range filter.

This is especially useful when trying to download a dataset from Numetric that has more than 10,000 rows.

This function uses the V2 version of the API.
}

