% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTables.R
\name{getTables}
\alias{getTables}
\title{Get a summary list of tables available to the API user (V3)}
\usage{
getTables(apiKey)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}
}
\description{
Returns a table with meta-data about all the tables. This function uses the V3 version of the API.
}
