% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDataset.R
\name{updateDataset}
\alias{updateDataset}
\title{Updates the parameters of an existing dataset (V2)}
\usage{
updateDataset(apiKey, datasetId, numetricName, dataframeName,
  category = "New Data", autocompletes = "", geoshapes = "",
  geopoints = "", boolean = "")
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{datasetId}{The dataset ID, in quotes, to be updated. It can be found by using the getDatasets function, or by navigating to the dataset in Numetric, and selecting the string after the last forward slash.}

\item{numetricName}{The name, in quotes, given to the dataset in the Numetric Warehouse.}

\item{dataframeName}{The name, without quotes, of the dataframe that will be used to create the dataset in Numetric.}

\item{category}{The category, in quotes, given to the dataset in the Numetric Warehouse.}

\item{autocompletes}{The column name, or vector of column names, in quotes, that will autocomplete in the Numetric quick search bar.}

\item{geoshapes}{The column name, or vector of column names, in quotes, that will be stored as a geoShape (point on a map). This needs to be in the format, "lat,long".}

\item{geopoints}{The column name, or vector of column names, in quotes, that will be stored as a geoPoint (for heatmaps). This needs to be in the format, "lat,long".}

\item{boolean}{The column name, or vector of column names, in quotes, that will be stored as True/False. True = 1, False = 0.}

\item{id}{The name of the column, in quotes, that will be used as the primary key. Each row should have a unique id, otherwise, the last row indexed will be the one saved.}
}
\value{
Returns the status of the update.
}
\description{
Uupdate an existing dataset in a Numetric Org.

This is useful when you want to add new columns to a dataset. You can also use it change the name or category of a dataset.

This function uses the V2 version of the API.
}

