% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllRows.R
\name{getAllRows}
\alias{getAllRows}
\title{Get up to 10,000 rows of a dataset (V2)}
\usage{
getAllRows(apiKey, datasetId)
}
\arguments{
\item{apiKey}{The API key, in quotes. You can find the API key in the settings after logging into Numetric}

\item{bucketVar}{The name of the dataset column, in quotes, that will be used to bucket the data.}
}
\value{
Returns a dataframe.
}
\description{
Returns a dataframe with up to 10,000 rows of a dataset.

This function uses the V2 version of the API.
}

