% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetsV3.R
\name{getDatasetsV3}
\alias{getDatasetsV3}
\title{Get a table of datasets (V3)}
\usage{
getDatasetsV3(apiKey)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}
}
\description{
Returns a table with meta-data about all the tables. This function uses the V3 version of the API.
}
