% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDashboard.R
\name{updateDashboard}
\alias{updateDashboard}
\title{Updates an existing Numetric dashboard (V2)}
\usage{
updateDashboard(apiKey, datasetId, name = "None", category = "None",
  content)
}
\arguments{
\item{apiKey}{You can find the API key in the settings after logging into Numetric}

\item{name}{The name of the dashboard, in quotes, in Numetric. Defaults to "None".}

\item{category}{The category, in quotes, of the dashboard. Defaults to "None".}

\item{content}{JSON formatted string of text. See https://nrl.readme.io/docs for more information.}
}
\description{
Updates the name, category, and content of an existing Numetric dashboard.

This function uses the V2 version of the API.
}
